Revert "#1072 refactor all editor tools to classes and ts (#1075)"
This reverts commit bacda2709423226343b1dea7646cad9d4489969e.
