+++ packages/ketcher-react/src/script/editor/Editor.ts
@@ -186,8 +186,30 @@ class Editor implements KetcherEditor {
       this._tool.cancel()
     }
 
-    let tool = new toolMap[name](this, opts)
-
+    // TODO: when all tools are refactored to classes, remove this check
+    // and use new keyword for every tool
+    let tool
+    const toolsAsClasses = [
+      'select',
+      'eraser',
+      'bond',
+      'chain',
+      'charge',
+      'rotate',
+      'atom',
+      'apoint',
+      'template',
+      'sgroup',
+      'rgroupfragment',
+      'rgroupatom',
+      'paste',
+      'attach'
+    ]
+    if (toolsAsClasses.includes(name)) {
+      tool = new toolMap[name](this, opts)
+    } else {
+      tool = toolMap[name](this, opts)
+    }
     if (!tool) {
       return null
     }
