+++ packages/ketcher-react/src/script/editor/tool/enhanced-stereo.ts
@@ -23,11 +23,13 @@ import {
 
 import Editor from '../Editor'
 
-class EnhancedStereoTool {
-  editor: Editor | undefined
-
-  constructor(editor) {
+function EnhancedStereoTool(
+  this: typeof EnhancedStereoTool,
+  editor: Editor
+): null | void {
+  if (!(this instanceof EnhancedStereoTool)) {
     const selection = editor.selection()
+
     const stereoAtoms = findStereoAtoms(
       editor.struct(),
       selection
@@ -35,11 +37,11 @@ class EnhancedStereoTool {
         : Array.from(editor.struct().atoms.keys())
     )
 
-    if (stereoAtoms.length !== 0) {
-      changeAtomsStereoAction(editor, stereoAtoms).then(
-        action => action && editor.update(action)
-      )
-    }
+    if (stereoAtoms.length === 0) return null
+
+    changeAtomsStereoAction(editor, stereoAtoms).then(
+      action => action && editor.update(action)
+    )
   }
 }
